refactor: notification 위치 변경
