+++ src/main/java/devut/buzzerbidder/domain/notification/controller/NotificationSseController.java
@@ -1,12 +1,11 @@
-package devut.buzzerbidder.global.notification.controller;
+package devut.buzzerbidder.domain.notification.controller;
 
 import devut.buzzerbidder.domain.user.entity.User;
-import devut.buzzerbidder.global.notification.service.SseEmitterPool;
+import devut.buzzerbidder.domain.notification.service.SseEmitterPool;
 import devut.buzzerbidder.global.requestcontext.RequestContext;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
