+++ src/main/java/devut/buzzerbidder/domain/notification/dto/NotificationListResponse.java
@@ -1,4 +1,4 @@
-package devut.buzzerbidder.global.notification.dto;
+package devut.buzzerbidder.domain.notification.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
